# app/frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionConfirmationRequest": {
        "additionalProperties": false,
        "description": "Confirm or reject one prepared action.\n\nConfirmation is a distinct API call with a closed vocabulary. A user typing\n\"okay\" into the chat endpoint does not confirm anything: `/api/chat` has no\npath to execution at all, which is what makes the gate structural rather\nthan a matter of parsing intent correctly.",
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ConfirmationDecision"
          },
          "expected_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Fingerprint"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ActionConfirmationRequest",
        "type": "object"
      },
      "ActionConfirmationResponse": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ExecutedActionView"
          },
          "action_status": {
            "$ref": "#/components/schemas/ActionState"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "action_status",
          "action",
          "message"
        ],
        "title": "ActionConfirmationResponse",
        "type": "object"
      },
      "ActionDetailResponse": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ExecutedActionView"
          },
          "action_status": {
            "$ref": "#/components/schemas/ActionState"
          }
        },
        "required": [
          "action_status",
          "action"
        ],
        "title": "ActionDetailResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "additionalProperties": false,
        "description": "A natural-language request plus the identity it is made under.\n\n`user_id` establishes the authorization context; it is looked up in the\nmock principal directory server-side. It is *not* trusted to describe its\nown permissions, and nothing in `message` can alter them \u2014\nsee `app/backend/auth/principals.py`.",
        "properties": {
          "account_scope": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Scope"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "description": "The full result of one natural-language request.",
        "properties": {
          "account_scope": {
            "items": {
              "type": "string"
            },
            "title": "Account Scope",
            "type": "array"
          },
          "action_status": {
            "$ref": "#/components/schemas/ActionState",
            "default": "none"
          },
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "escalation_recommended": {
            "default": false,
            "title": "Escalation Recommended",
            "type": "boolean"
          },
          "outcome": {
            "$ref": "#/components/schemas/ResponseOutcome"
          },
          "policy_decisions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PolicyDecisionView"
            },
            "title": "Policy Decisions",
            "type": "array"
          },
          "proposed_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposedActionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "reference_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Time"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "responded_at_utc": {
            "format": "date-time",
            "title": "Responded At Utc",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "sources": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "title": "Sources",
            "type": "array"
          },
          "step_budget_exhausted": {
            "default": false,
            "title": "Step Budget Exhausted",
            "type": "boolean"
          },
          "tools_used": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ToolUse"
            },
            "title": "Tools Used",
            "type": "array"
          },
          "uncertainties": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Uncertainties",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "session_id",
          "outcome",
          "answer",
          "user_id",
          "role",
          "account_scope",
          "responded_at_utc"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PendingActionsResponse": {
        "properties": {
          "actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProposedActionView"
            },
            "title": "Actions",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "PendingActionsResponse",
        "type": "object"
      },
...
      "ProposedActionView": {
        "description": "A prepared action awaiting a human. Nothing here has happened yet.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "confirmation_required": {
            "default": true,
            "title": "Confirmation Required",
            "type": "boolean"
          },
          "evidence_chunk_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Evidence Chunk Ids",
            "type": "array"
          },
          "expires_at_utc": {
            "format": "date-time",
            "title": "Expires At Utc",
            "type": "string"
          },
          "parameter_fingerprint": {
            "title": "Parameter Fingerprint",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Parameters",
            "type": "object"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/ActionState"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "action_type",
          "status",
          "target_type",
          "target_id",
          "account_id",
          "parameters",
          "preview",
          "reason",
          "expires_at_utc",
          "parameter_fingerprint"
        ],
        "title": "ProposedActionView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Natural-language support and operations assistant over ParcelPilot's policy\npack and operational records.\n\nThe agent reasons; deterministic code decides. Account scoping, source\nprecedence, cancellation and service-credit arithmetic, and every state change\nare enforced below the model layer and cannot be influenced by the contents of\na request.\n\nState-changing actions are prepared by `POST /api/chat` and executed only by\n`POST /api/actions/{action_id}/confirm`. Nothing in a chat message can perform\none.\n",
    "title": "ParcelPilot Support & Operations Agent API",
    "version": "0.8.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/actions/pending": {
      "get": {
        "description": "Proposals awaiting confirmation, within the caller's account scope.",
        "operationId": "pending_actions_api_actions_pending_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "header",
            "name": "x-parcelpilot-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parcelpilot-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pending Actions",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/actions/{action_id}": {
      "get": {
        "description": "The audit record for one action: state, timeline, and its effect.\n\nScoped like every other read \u2014 an action on another customer's account is\nreported as absent, not as forbidden, for the same reason the record\nlayer does.",
        "operationId": "action_detail_api_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "header",
            "name": "x-parcelpilot-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parcelpilot-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Detail",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/actions/{action_id}/confirm": {
      "post": {
        "description": "Execute or reject one prepared action. The only path to a state change.\n\nEvery check the Phase 4 state machine offers is applied, under the\n*confirming* caller rather than the preparing one: the action must exist\nwithin this caller's account scope, belong to this conversation, still be\npending, not have expired, still have a live target, and \u2014 when a\nfingerprint is supplied \u2014 still describe exactly what was reviewed. The\nstatus guard inside the UPDATE makes execution single-use even if two\nconfirmations race.",
        "operationId": "confirm_action_api_actions__action_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-parcelpilot-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parcelpilot-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionConfirmationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Action",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/api/chat": {
      "post": {
        "description": "Answer one natural-language request with its evidence.\n\nThe authorization context is built from the caller's identity before the\nmessage is read, and nothing in the message can widen it. Account ids in\nprose are treated as text; the agent's scope comes from records it was\npermitted to resolve.",
        "operationId": "chat_api_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-parcelpilot-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parcelpilot-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}